fix: update several test suites to work with Stacks 2.1 docker image
